Nasa_example/tap-nasa-incremental/requirements.txt
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.11.1
singer-python==5.9.0